move qwen2_vl test into test_vision_openai_server.py
