+++ test/srt/run_suite.py
@@ -28,7 +28,6 @@ suites = {
         "test_triton_attn_backend.py",
         "test_update_weights.py",
         "test_vision_openai_server.py",
-        # "test_qwen2_vl_openai_server.py", # TODO: remove this comment
     ],
     "sampling/penaltylib": glob.glob(
         "sampling/penaltylib/**/test_*.py", recursive=True
